+++ packages/plugin-testing/jest.config.js
@@ -1,0 +1,8 @@
+const base = require('../../jest.config.base.js');
+const packageJson = require('./package.json');
+
+module.exports = {
+  ...base,
+  name: packageJson.name,
+  displayName: packageJson.name,
+};