+++ CHANGELOG.md
@@ -3,6 +3,27 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.4.0](https://github.com/iterativelyhq/itly-sdk/compare/v1.3.0...v1.4.0) (2020-10-28)
+
+
+### Bug Fixes
+
+* fixed linter error & CR feedback ([b3ca320](https://github.com/iterativelyhq/itly-sdk/commit/b3ca3208179fbf2065670d34728bad31010a54a3))
+* made LOGGERS a frozen readonly object ([71ff140](https://github.com/iterativelyhq/itly-sdk/commit/71ff140ba17818944fe2d01635647aec0fa2ed8a))
+* plugin-testing/node - remove userId from methods return results ([2c7bee5](https://github.com/iterativelyhq/itly-sdk/commit/2c7bee5b81d3e38bb60b6ec1750e81b889b4c823))
+* set correct environment for plugin-testing tests ([0cb0618](https://github.com/iterativelyhq/itly-sdk/commit/0cb06189186467559778c9de38f62e320e0aa353))
+* wrapped individual plugin method calls in try catch ([932417c](https://github.com/iterativelyhq/itly-sdk/commit/932417c4630e5d51340e8af03bbe50c00b710a74))
+
+
+### Features
+
+* added plugin-testing ([43e65bc](https://github.com/iterativelyhq/itly-sdk/commit/43e65bca097f50b770d1f2277f036dc02dd691eb))
+* added plugin-testing ([fb411bb](https://github.com/iterativelyhq/itly-sdk/commit/fb411bb66f5d371e9a32cc3187c1158e68a59543))
+
+
+
+
+
 # [1.3.0](https://github.com/iterativelyhq/itly-sdk/compare/v1.2.0...v1.3.0) (2020-09-08)